Sketch some project structure and some checkpoints to aim towards
I think I do want to implement, at least to some degree, an
end-to-end compiler from the start. This should help make sure that
things are working.

This was my strategy before, so I know it isn't a perfect one. I think
the other technique I will use is: simply don't care too much about
type-system perfection, and the like. I will be able to be more
productive if I can have some information available on the syntax
tree that has to be filled in over time, and is in an invalid state
sometimes. Good coding practice can help make this work better.