Start leaking lexemes; An interesting codegen idea
It generally makes sense to leak the lexemes because they should live
for the whole program anyways. I suppose if we ever want some kind
of longer-running program we may not want to... but even then, there
will still be pretty much re-use in terms of lexemes across re-invocations,
so it actually kinda makes sense to keep them around.

TBH, probably the best approach for that longer-running program would
be to keep the state around for a while and then occasionally start
from scratch.