Implement basic print statement.. doesn't support num constants yet
This is really getting to the heart of ... how do we want to type
check the numerical constants.

The 'context' system sort of works, although it doesn't work perfectly.
We could also make the type checker propagate these types back in,
but that is a bit slow.

I suppose there could be a second type checking pass to resolve
the numerical types, which wouldn't be too bad, as the type checking
is currently more or less the fastest stage on the big benchmarks.

But, the 'context' system has the advantage that it's parallelizable.

On the other hand... keeping codegen as simple as possible may be the
best strategy.