Implement partially working underlining
This almost works, but it definitely has some off-by-one somewhere.

The question is whether the SourceLocations are being generated
correctly.

From what I can tell, for try_return_float_for_int.poni, the correct
offset for the return should be 35, but the program is reporting 38...
which is an off-by-3-error....