Add promotion rules to the rest of codegen.rs -- add some tests
As expected, because the blocks in typecheck.rs do not ever
call promote(), promote_to_float_block_assign_assumeinit.poni fails
with a panic in codegen.rs.

This should be fixable in typecheck.rs, I think. I think fixing it
in codegen.rs would be wrong, due to it not being able to traverse
far enough down the tree just through promote() in codegen.