Add some tests to the tester
To help make sure at least part of our promotion logic is working.

That said, we need function calls to get the rest working.

Also, we could consider building some testing infrastructure
that is less rough-around-the-edges than this -- right now it is
heavily dependent on output format, which is annoying. It might
be better to have something like

    expect_int(5);
    expect_float(10);
    expect_str("meow");

and then have these function calls generated somehow, then have
the process simply indicate success/failure with an exit code.

(Basically, the 'print' function would be stubbed out for something
like ps_test_report_int() which would then push the results on to
a linked list or similar)