Add str_of_strbuf test -- fails in a strange way
When I run this test manually, it fails by just printing "hello".
But, the version compiled by the tester seems to work just fine...

Oh wait, it is the std=c11?

Edit: No, it was not std=c11 -- it was that I was running an
outdated copy of poniescript that was stil generate
ps_strfmt_str(..., strbuf->buffer).