Implement proper cname mangling, and test block_shadow.poni
Note that the cname mangling is apparently not STRICTLY necessary
just for shadowed locals -- it appears C is capable of handling that
itself -- but this helps us make sure the compiler is working as
expected.

Also note that now the globals tests are failing -- this is because
global variables are not able to be resolved without the binding
pass. This is actually expected behavior.