Add string types and some string type promotions
Now str_types.poni passes, including with a Str.

I suppose it might be bad to have these implicit promotions... but
on the other hand, it is very convenient to be able to allocate a
string based on a literal, and I think it is relatively clear that
that's what's happening, especially for StrBuf.