Essentially implement function calls
These bring with them a couple of concerns and performance considerations.

First, it is a bit annoying that the current plan is to have at
least two different AST nodes for calls--those being direct
function calls, as well as calls through fun values. But, that
might be fine.

Second, because the function has a Vec of params, there are some
considerations with passing that around, especially when stealing
it from an UnboundCall and putting it into an actual Call.

Similarly, this currently requires allocating a Vec inside the codegen
for storing all of the PromotedVals.