New C-level structure for function values
The idea here is that each function object is just a normal reference
type, like any other in PonieScript. (This does mean these generated
objects will eventually need a gc_object field).

The point is that it reduces the number of unrelated objects created
for a single closure, while also hopefully allowing the C compiler to
see certain optimizations for non-closure functions.