Add a basic "this" initializer to functions
This seems reasonable, but maybe not quite what we want.

We definitely want the "closure" parameter to just be the object,
if there is one -- and then we can implement closures in terms of
classes. But it's not clear how we want to implement "bound calling"
or whatever to make this clean.