Start working on nested gets: Looks like type inference will be something
So the problem here is that we can't meaningfully infer the type
of an expression until we have visited the class. In the short
term, there may be a partial fix for some of our tests in case
we're missing some type inference code somewhere, but that won't
help if the usage of the type comes before the class definition.

I think we might need some sort of way to infer types for a class
"right now" that we can call when we are accessing properties of
that class. Then we of course need some kind of graph thingy
to make sure that we don't recurse forever.