tests: implement basic //? for tests that should error.
At some point, we may wish to implement specific matching for errors,
because tests often should fail on a specific error. But anything is
better than nothing, and once this is done we can always grep for
//? to find all the tests to change later.

(Or, of course, they will just fail if we implement extra stuff that
they have to do, so we will just have to update them at that time).