Convert codegen of 'compiler-error' into panics -- esp. for Val::Bottom
This really just seems much better. It means if we ever get one of
these, instead of trying to figure out why, we can instantly get
a stacktrace... Maybe that's not great, but it seems better than
generating invalid C code.

This seems especially good for the Val::Bottom and PromotedVal::Bottom
cases, as those have caused a lot of issues historically.