New 'poni_hot.h' API for hot reloading
Now there is a bit of an API for making applications hot reload
poniescript.

The main problem is that dlopen() the same path straight up does
not work, while dlclose() BEFORE we dlopen() is liable to explode
(e.g. bus error).

I think the solution is maybe to provide a set of temporary paths
that are rotated through?