Maybe less-failed optimization?
The problem with the previous one was (probably) that we were still
creating a new string object, and actually doing it every time,
with std::mem::take.

Now we only create a new object when necessary, and it's not a copy
of an existing string.

Still, at least for big4.poni, this likely isn't a huge improvement
(if any), because the strings are pretty small anyway, so the
difference is kind of a wash.