More testing for the vec-of-vecs... it's unclear where the slowdown could be coming from.
I simply can't believe this isn't the right way to do this. Maybe
it will, somehow, be slower for our use case, but even then I
really can't see how that could be the case.

I would think the numbers don't lie, but I can't actually track down
what specifically is causing the vec-of-vecs version to be slower
than the other version. Without some specific reason to believe that
it is definitely better, I think I want to go with this version for
now, especially because it also seems more reasonable in terms of
future directions for expanding the local allocator state.