gc: Logging support--note this slows down link times HUGELY
I believe most of the slowdown is likely cause I also added env_logger,
which really shouldn't directly be integrated into the GC but it is
for now.

Of course, for hot reloading, this shouldn't be tooo bad because
everything should be dynamically linked anyway, so the GC will
already be linked together and such, but it is a bit sad.