Partial implementation of Nil syntax -- my mistake was Option<Unassigned>
I think Option[Unassigned] doesn't work here, at least not right now,
because it doesn't seem like a type that should be promoted from
unassigned. Although, this actually might just be an issue with
promote_from_unassigned, and we might want to check if a similar
issue doesn't apply to tuples.