"Fix" to else_ret_assigns
So the problem here was that my test code wasn't testing what
I thought it was, and the existance of the Horse?? type that kept
breaking was legit.

Essentially, because I am doing:

x: Horse? = maybe_make_Horse() else ...

the left hand side ends up promoted to Horse??, so that it can
unwrap and create Horse?.

To be honest, though, what I should probably do instead is make
it so that (x else ...) promoted to the non-option type just
directly realizes what is happening and allows that promotion
to occur...