ps_int: Change to int64_t?
I had sort of been thinking ps_int is int64_t (this is based on the
Godot int which is 64-bit).

This is very convenient for, at least, day 2 of Advent of Code.
But I wonder if we should implement additional integer types
instead...

I do think that 64 bit integers are useful to have in the
programming language, though, and that if you're only going to
have one int type, 64 bit is arguably better than 32. But I
guess I'll have to see what I want to do.