Add a test file that needs to be reduced to a minimum reproducible example
I ran into this panicking typecheck when working on advent of code.
I thiiink it is related to nested ifs in some way; the issue seems
to be the last statement of the 'if halfpow * 2 == pow' block.

Adding a dummy 0; to the end of that block gets rid of the panic,
for example.