DynArray: Some initial infrastructure
Dynamic arrays, as it turns out, are going to be pretty complicated.

The main issue is that we really want the backing buffer to be
'possibly uninitialized' because we really can't initialize
the elements that are e.g. nonnullable without a real value.

But this means that for memory safety purposes, the length, when
read, must *never* *overestimate* the number of elements that are
safe to read from the array. Fundamentally this means that when
populating a possibly-uninitialized slot, we will NEED the length
value to be updated *after* the slot itself, which could mean
using an atomic store for the length update.

It will be a bit tricky.

The good news is there are not many issues in a single-threaded
context. The bad news is that the garbage collector is concurrent.