ABI: Another test case in context-abi-comparison
Also, I forgot the sort of catch-22 we're in, so I wrote that down
as well.

(The catch-22: We need the GC frames stack for correctness, UNLESS
we are knowingly compiling to a context that can regularly safepoint,
e.g. a game loop that can have a top-level safepoint.

This could indicate that we want two different ABIs, but it's not
really clear).