rust-rt: Some basic Array methods
These would be more for usage from the Rust-side.

Note that it's unclear to me exactly how I want all this to work.

In particular, PsArray<> is a weird type write now. It's sort
of like the PonieScript array types, where we know what they are
at run-time; except this one is even weirder because it is
exclusively a *pointer* to such a type.

I guess we could make PsArray<> some sort of janky non-sized type,
and then make Rust users write Gp<PsArray<T>> or something like
that.